fixed script
