+++ docgen.cjs
@@ -4,6 +4,41 @@ const fs = require("fs")
 const path = require("path")
 const dmd = require("dmd")
 
+// First of all, we need to make a hack to the code to make it work with jsdoc2md
+// find and replace all strings "// @type definition" with "1 // @type definition"
+// this is a workaround for a bug in the babel plugin and jsdoc
+const ORIGINAL = "// @type definition"
+const REPLACEMENT = "1 // WORKARAOUND"
+
+const dir = path.join(process.cwd(), "src")
+
+const replaceToDir = (dir, reverse = false) => {
+    const files = fs.readdirSync(dir)
+    for (const file of files) {
+        // check if file is a directory
+        if (fs.statSync(path.join(dir, file)).isDirectory()) {
+            replaceToDir(path.join(dir, file), reverse)
+            continue
+        }
+
+        if (!file.endsWith(".ts")) {
+            continue
+        }
+        const filePath = path.join(dir, file)
+        const content = fs.readFileSync(filePath, "utf8")
+        if (!reverse) {
+            // replace all occurrences of ORIGINAL with REPLACEMENT
+            const newContent = content.replace(new RegExp(ORIGINAL, "g"), REPLACEMENT)
+            fs.writeFileSync(filePath, newContent)
+        } else {
+            const newContent = content.replace(new RegExp(REPLACEMENT, "g"), ORIGINAL)
+            fs.writeFileSync(filePath, newContent)
+        }
+    }
+}
+
+replaceToDir(dir)
+
 const docsFolder = "docs"
 
 // This is where the root plugin folder is, relative to this script
@@ -51,3 +86,6 @@ const links = plugins.map(({ pluginName }) => {
 })
 
 fs.writeFileSync(path.join(__dirname, "DOCS.md"), header + links.join(""))
+
+// Restore the original code
+replaceToDir(dir, true)
